# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Med Fusion OS API
  description: >
    API da Ordem de Serviço da Med Fusion Manutenção e Venda
    Clínica Hospitalar Ltda. Fecha a issue #21 (Definir contrato da API e
    versionamento /api/v1). Escopo e critérios de aceite em
    https://github.com/celsojuniorsfs/medfusion-os-web/blob/main/docs/escopo-v1.md.
    Convenções (autenticação, formato de erro) detalhadas em ./api-conventions.md.

    Atualizado em 07/09/2026 (duas rodadas de validação com o cliente): equipamentos sem limite
    por OS (catálogo reaproveitável por cliente), peça com valor opcional + labor_cost, status
    ampliado (OrderStatus, com not_approved distinto de canceled) e notificação automática por
    e-mail/WhatsApp na criação da OS.

    Atualizado em 08/09/2026: nomes de campo (schemas) padronizados em inglês — só os valores de
    negócio em português (nomes de cliente, defeitos descritos etc.) continuam em português,
    como dados, não como nomes de campo.

    Atualizado em 08/09/2026 (arquitetura): identificadores (`id` e os parâmetros de rota
    correspondentes) passaram de inteiro sequencial para UUID — identidade dos agregados do
    Event Sourcing (ver architecture.md). `number` da OS não muda: continua o número de negócio
    sequencial (seed 1336), editável pelo técnico, sem relação com a identidade do agregado.
  version: "2.0.0"
  license:
    name: Proprietary
    identifier: proprietary
servers:
  - url: http://localhost:8000/api/v1
...
security:
  - bearerAuth: []
paths:
...
  /orders:
    get:
      tags: [orders]
      summary: Lista OS com filtros, ordenadas por status (cancelada por último) e depois data decrescente
      operationId: listOrders
      parameters:
        - name: client_id
          in: query
          schema: { type: string, format: uuid }
        - name: equipment_id
          in: query
          description: >
            Filtra as OS's de um equipamento específico do catálogo. Usado pela tela de
            histórico do equipamento (ver escopo-v1.md § Histórico do equipamento) — geralmente
            combinado com client_id.
          schema: { type: string, format: uuid }
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/OrderStatus" }
        - name: date_from
          in: query
          schema: { type: string, format: date }
        - name: date_to
          in: query
          schema: { type: string, format: date }
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
      responses:
        "200":
          description: Página de OS
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Pagination"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/Order" }
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [orders]
      summary: Cria uma OS com equipamentos e peças aninhados
      description: >
        Equipamentos informados sem equipment_id são cadastrados no catálogo do cliente nesta
        mesma chamada (ver OrderEquipmentInput). O PDF não é gerado automaticamente — o técnico
        solicita via POST /orders/{id}/pdf (ver abaixo). A geração automática + notificação por
        e-mail/WhatsApp na criação chega com o api#70, que ainda depende da verificação da conta
        Meta do cliente.
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrderInput" }
      responses:
        "201":
          description: OS criada
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Order" }
        "422":
          $ref: "#/components/responses/ValidationError"
        "409":
          description: Número de OS já utilizado por outra ordem de serviço
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Sanctum personal access token, emitido por POST /auth/login
  parameters:
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PerPage:
      name: per_page
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 15 }
  responses:
    Unauthorized:
      description: Token ausente, inválido ou expirado
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    ValidationError:
      description: Erro de validação
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ValidationErrorBody" }
  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message: { type: string }
    ValidationErrorBody:
      type: object
      required: [message, errors]
      properties:
        message: { type: string, example: "The given data was invalid." }
        errors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
          example:
            number: ["O número informado já está em uso."]
    Pagination:
      type: object
      properties:
        links:
          type: object
          properties:
            first: { type: string, format: uri }
            last: { type: string, format: uri }
            prev: { type: [string, "null"], format: uri }
            next: { type: [string, "null"], format: uri }
        meta:
          type: object
          properties:
            current_page: { type: integer }
            last_page: { type: integer }
            per_page: { type: integer }
            total: { type: integer }
    User:
      type: object
      description: >
        role (api#133) é uma fatia mínima pro sistema de alertas mirar quem recebe cada aviso e
        liberar o liga/desliga só pro admin geral — não é o RBAC completo do futuro perfil
        Financeiro. Nenhum endpoint existente passa a checar isso.
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        email: { type: string, format: email }
        role:
          {
            type: string,
            enum: [technician, administrative, general_admin],
            description: "Técnico (nunca recebe alerta), administrativo da assistência, ou administrador geral (único que liga/desliga notificação)",
          }
    Client:
      type: object
      description: >
        Nomes de campo em inglês (decidido em 08/09/2026); os valores em si continuam em
        português — são dados de negócio do cliente, não nomenclatura do sistema. A partir de
        10/09/2026 cobre pessoa física e jurídica explicitamente (ver person_type) — antes disso
        o cadastro só previa pessoa jurídica.
      properties:
        id: { type: string, format: uuid }
        person_type:
          {
            type: string,
            enum: [individual, company],
            description: "Pessoa física (individual) ou jurídica (company)",
          }
        name: { type: string, description: "Nome completo (pessoa física) ou razão social (pessoa jurídica)" }
        trade_name: { type: ["string", "null"], description: "Nome fantasia (só pessoa jurídica)" }
        tax_id:
          {
            type: string,
            description: "CPF (pessoa física, 11 dígitos) ou CNPJ (pessoa jurídica, 14 dígitos) — só números",
            example: "31233218000110",
          }
        state_registration: { type: ["string", "null"], description: "Inscrição estadual (só pessoa jurídica)" }
        requester: { type: ["string", "null"], description: "Solicitante" }
        department: { type: ["string", "null"], description: "Setor" }
        phone: { type: ["string", "null"], description: "Telefone" }
        email: { type: ["string", "null"], format: email, description: "E-mail" }
        address: { type: ["string", "null"], description: "Endereço" }
        city: { type: ["string", "null"], description: "Cidade" }
        state: { $ref: "#/components/schemas/BrazilianState" }
        postal_code: { type: ["string", "null"], description: "CEP, só números", example: "15775000" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    BrazilianState:
      type: ["string", "null"]
      description: UF (sigla do estado)
      enum:
        [
          null, AC, AL, AP, AM, BA, CE, DF, ES, GO, MA, MT, MS, MG, PA, PB, PR, PE, PI, RJ, RN,
          RS, RO, RR, SC, SP, SE, TO,
        ]
...
    OrderStatus:
      type: string
      description: >
        Fechado nas duas rodadas de validação de 07/09/2026 (nomes de valor traduzidos para
        inglês em 08/09/2026 — o significado de negócio não mudou). Fluxo: open → in_analysis →
        external_quote (opcional) → awaiting_approval → approved → completed. canceled é
        alcançável a partir de qualquer estado anterior a approved. not_approved só a partir de
        awaiting_approval — orçamento sem retorno do cliente, mudança manual do técnico;
        distinto de canceled (que é uma decisão explícita do cliente ou da empresa).
        warranty_repair é uma reabertura especial, só a partir de completed (retrabalho dentro
        do prazo de garantia, mesma OS) — depois volta a completed. partially_completed (api#140)
        deriva sozinho da situação de cada equipamento (ver OrderEquipmentSnapshot.situation) —
        nunca alcançável por PATCH /orders/{id}/status manual, mesmo estando na tabela de
        transições. awaiting_approval → approved (api#149) tem os dois caminhos: manual (a OS
        inteira, como sempre foi) E automático, assim que o primeiro equipamento é aprovado (ver
        OrderEquipmentSnapshot.approval_status) — os dois convivem, um não bloqueia o outro. Ver
        api-conventions.md para a tabela completa e a correspondência com os termos em português
        usados com o cliente (aberta, em análise, orçamento externo, aguardando aprovação,
        aprovada, não aprovado, parcialmente concluída, garantia, concluída, cancelada).
      enum:
        - open
        - in_analysis
        - external_quote
        - awaiting_approval
        - approved
        - not_approved
        - partially_completed
        - warranty_repair
        - completed
        - canceled
...
    OrderEquipmentAccessory:
      type: object
      description: >
        Acessório digitado pra este equipamento nesta OS — snapshot livre, sem vínculo com o
        catálogo estruturado de acessórios (Accessory/EquipmentAccessory).
      required: [name, quantity]
      properties:
        name: { type: string, maxLength: 255 }
        quantity: { type: integer, minimum: 1 }
...
    OrderEquipmentInput:
      description: >
        Referencia um equipamento já cadastrado (equipment_id) OU cadastra um novo pelo nome —
        `OrderRequest::rules()` expressa esse oneOf com `required_without` nas duas pontas, já que
        o Laravel não tem oneOf nativo.

        `accessories` aqui é uma LISTA de OrderEquipmentAccessory (`{name, quantity}`) e vale nos
        dois ramos — é o snapshot desta OS (`order_equipment_accessories`), sem vínculo com o
        catálogo estruturado de acessórios do equipamento (api#92). Até 26/09/2026 era texto livre
        (string, até 255 caracteres); um cliente antigo que ainda mandar string aqui é aceito por
        um shim transitório em `OrderRequest::prepareForValidation()` e convertido pra lista
        (dividida por vírgula/ponto-e-vírgula) antes da validação — não confiar nesse formato pra
        payloads novos, ele será removido depois que todos os clientes migrarem.

        Corrigido em 21/09/2026: o oneOf anterior referenciava `EquipmentInput` no segundo ramo,
        que desde o api#112 exige `equipment_model_id`/`no_accessories`/acessórios estruturados —
        um payload que `POST /orders` nunca aceitou. O contrato tinha ficado pra trás depois dessa
        mudança no catálogo.

        `preventive_maintenance`/`calibration` (api#146) valem nos dois ramos, pelo mesmo motivo
        de `accessories` — corrigido do nível de OS (api#134) pro de equipamento, porque nem todo
        equipamento tem calibração.

        `labor_cost`/`items` (api#149) também valem nos dois ramos — mão de obra e peças por
        equipamento, além dos campos gerais de mesmo nome no nível raiz de OrderInput (sem
        vínculo com nenhum equipamento).
      # Sem os `not`, um payload com equipment_id E name ao mesmo tempo batia nos dois ramos
      # (nenhum dos dois usa additionalProperties:false, então a chave do OUTRO ramo é só
      # "propriedade extra permitida") — um validador de contrato estrito rejeitava algo que a API
      # aceita de verdade (prioriza equipment_id, ignora o resto). O `not` força só um ramo bater
      # por vez.
      oneOf:
        - type: object
          required: [equipment_id]
          not: { required: [name] }
          properties:
            equipment_id: { type: string, format: uuid }
            preventive_maintenance: { type: boolean, default: false }
            calibration: { type: boolean, default: false }
            labor_cost: { type: ["number", "null"], format: float }
            accessories:
              type: array
              items: { $ref: "#/components/schemas/OrderEquipmentAccessory" }
            items:
              type: array
              items: { $ref: "#/components/schemas/OrderItem" }
        - type: object
          required: [name]
          not: { required: [equipment_id] }
          properties:
            name: { type: string, maxLength: 255 }
            brand: { type: ["string", "null"], maxLength: 255 }
            model: { type: ["string", "null"], maxLength: 255 }
            serial_number: { type: ["string", "null"], maxLength: 255 }
            asset_tag: { type: ["string", "null"], maxLength: 255 }
            preventive_maintenance: { type: boolean, default: false }
            calibration: { type: boolean, default: false }
            labor_cost: { type: ["number", "null"], format: float }
            accessories:
              type: array
              items: { $ref: "#/components/schemas/OrderEquipmentAccessory" }
            items:
              type: array
              items: { $ref: "#/components/schemas/OrderItem" }
    OrderEquipmentSituation:
      type: string
      description: >
        Situação técnica de UM equipamento dentro da OS (api#140), independente do status da OS
        — sem máquina de estados, qualquer uma pode ir pra qualquer outra. `completed` e
        `returned_unrepaired` são as "resolvidas": derivam o status da OS (ver
        OrderStatus.description) e param a contagem pro alerta por situação (api#147).
      enum:
        - in_analysis
        - awaiting_part
        - external_repair
        - completed
        - returned_unrepaired
    OrderEquipmentApprovalStatus:
      type: string
      description: >
        Status do orçamento de UM equipamento (api#149), independente de OrderEquipmentSituation.
        Sem coluna correspondente no payload de entrada — nasce `null` (nenhum orçamento gerado
        ainda) e vira `awaiting_approval` sozinho no primeiro POST /orders/{id}/pdf que inclui o
        equipamento; só depois disso PATCH /orders/{id}/equipments/approval aceita mudar pra
        `approved`/`not_approved`.
      enum:
        - awaiting_approval
        - approved
        - not_approved
    OrderEquipmentSnapshot:
      type: object
      description: >
        Decisão da Fase 2 (Análise): retrato do equipamento no momento em que esta OS foi criada
        — NÃO reflete edições feitas depois no catálogo (`equipments`). `equipment_id` aponta
        para o cadastro atual (usado para navegar até ele e para o filtro de histórico,
        `GET /orders?equipment_id=X`), mas os demais campos são a cópia congelada.
      required: [accessories]
      properties:
        id:
          {
            type: string,
            format: uuid,
            description: "id de order_equipments — usado em PATCH /orders/{id}/equipments/situation (api#140)",
          }
        equipment_id: { type: ["string", "null"], format: uuid }
        name: { type: string }
        brand: { type: ["string", "null"] }
        model: { type: ["string", "null"] }
        serial_number: { type: ["string", "null"] }
        asset_tag: { type: ["string", "null"] }
        preventive_maintenance:
          {
            type: boolean,
            description: "Alimenta a contagem de 12 meses até a próxima revisão (api#136)",
          }
        calibration:
          {
            type: boolean,
            description: "Manutenção distinta da preventiva, mas sem vencimento próprio: sempre sai na mesma data/mesma OS",
          }
        situation: { $ref: "#/components/schemas/OrderEquipmentSituation" }
        situation_changed_at:
          {
            type: string,
            format: date-time,
            description: "Última mudança de situação — alimenta o alerta por situação (api#147)",
          }
        completed_at:
          {
            type: ["string", "null"],
            format: date-time,
            description: "Preenchida só enquanto situation = completed; limpa se voltar a ficar pendente (ex.: reabertura em garantia). Base da contagem de 12 meses (api#136).",
          }
        position:
          {
            type: integer,
            description: "Ordem de chegada dentro da OS — vira a letra do certificado (api#61: 0 = A, 1 = B...)",
          }
        approval_status:
          {
            allOf: [{ $ref: "#/components/schemas/OrderEquipmentApprovalStatus" }],
            nullable: true,
            description: "null = nenhum orçamento gerado ainda pra este equipamento (api#149)",
          }
        approval_status_changed_at: { type: ["string", "null"], format: date-time }
        labor_cost:
          {
            type: ["number", "null"],
            format: float,
            description: "Mão de obra específica deste equipamento — convive com orders.labor_cost (geral)",
          }
        items:
          type: array
          description: Peças vinculadas a este equipamento — itens sem vínculo ficam em Order.items.
          items: { $ref: "#/components/schemas/OrderItem" }
        accessories:
          type: array
          description: Sempre presente, mesmo vazia ([]).
          items: { $ref: "#/components/schemas/OrderEquipmentAccessory" }
    OrderItem:
      type: object
      description: >
        unit_price é opcional — corrigido na validação de 07/09/2026. Orçamentos de prefeitura
        costumam trazer só o valor da mão de obra (orders.labor_cost), com o valor das peças
        embutido, para não ultrapassar o teto que dispara licitação.
      required: [quantity, description]
      properties:
        quantity: { type: number }
        description: { type: string }
        unit_price: { type: ["number", "null"], format: float }
...
    Order:
      type: object
      description: >
        Nomes de campo em inglês (decidido em 08/09/2026); os checkboxes de tipo de atendimento
        e os campos de texto continuam guardando dados em português — o formulário e o
        vocabulário de negócio com o cliente não mudam, só o nome técnico do campo.
      properties:
        id: { type: string, format: uuid }
        number: { type: integer, example: 1336, description: "Número da OS" }
        date: { type: string, format: date }
        status: { $ref: "#/components/schemas/OrderStatus" }
        client: { $ref: "#/components/schemas/Client" }
        user: { $ref: "#/components/schemas/User" }
        picked_up: { type: boolean, description: "Retirado" }
        warranty:
          { type: boolean, description: "Garantia (tipo de atendimento)" }
        technical_training:
          { type: boolean, description: "Treinamento técnico" }
        on_site_quote: { type: boolean, description: "Orç. local" }
        rental: { type: boolean, description: "Locação" }
        reported_defect:
          { type: ["string", "null"], description: "Defeito apresentado" }
        maintenance_plan:
          { type: ["string", "null"], description: "Manutenção a aplicar" }
        notes: { type: ["string", "null"], description: "Observação" }
        payment_method:
          { type: ["string", "null"], description: "Forma de pagamento" }
        warranty_period:
          { type: ["string", "null"], description: "Garantia (prazo)" }
        proposal_validity:
          { type: ["string", "null"], description: "Validade da proposta" }
        labor_cost:
          {
            type: ["number", "null"],
            format: float,
            description: "Valor da mão de obra",
          }
        total: { type: number, format: float }
        certificate_number:
          { type: ["string", "null"], description: "Nº do certificado" }
        pdf_generated_at:
          {
            type: ["string", "null"],
            format: date-time,
            description: "Quando o PDF mais recente foi gerado (null se nunca foi). Ver GET /orders/{id}/pdf pra baixar",
          }
        equipments:
          type: array
          items: { $ref: "#/components/schemas/OrderEquipmentSnapshot" }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItem" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    OrderInput:
      type: object
      description: >
        total não é enviado pelo cliente — é sempre calculado no backend a partir de labor_cost,
        items e, desde a api#149, dos labor_cost/items de cada equipamento também. Pelo menos um
        valor precisa estar presente em ALGUM desses lugares (ver critérios de aceite em
        escopo-v1.md).
      required: [number, date, client_id, equipments, items]
      properties:
        number: { type: integer }
        date: { type: string, format: date }
        client_id: { type: string, format: uuid }
        picked_up: { type: boolean, default: false }
        warranty: { type: boolean, default: false }
        technical_training: { type: boolean, default: false }
        on_site_quote: { type: boolean, default: false }
        rental: { type: boolean, default: false }
        reported_defect: { type: ["string", "null"] }
        maintenance_plan: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        payment_method: { type: ["string", "null"] }
        warranty_period: { type: ["string", "null"] }
        proposal_validity: { type: ["string", "null"] }
        labor_cost: { type: ["number", "null"], format: float }
        equipments:
          type: array
          minItems: 1
          description: Sem limite de quantidade — corrigido na validação de 07/09/2026.
          items: { $ref: "#/components/schemas/OrderEquipmentInput" }
        items:
          type: array
          description: Gerais, sem vínculo com nenhum equipamento (api#149) — os demais vão em equipments[].items.
          items: { $ref: "#/components/schemas/OrderItem" }